+++ src/node/rpc/test/frontend_test.cpp
@@ -495,28 +495,15 @@ void prepare_callers(NetworkState& network)
 
   auto tx = network.tables->create_tx();
   network.tables->set_encryptor(encryptor);
+  
+  init_network(network);
 
   GenesisGenerator g(network, tx);
-  g.create_service();
+  g.create_service(network.identity->cert);
   user_id = g.add_user({user_caller});
   member_id = g.add_member(member_cert);
   invalid_member_id = g.add_member(invalid_caller);
   CHECK(tx.commit() == kv::CommitResult::SUCCESS);
-}
-
-void add_callers_bft_store()
-{
-  auto gen_tx = bft_network.tables->create_tx();
-  bft_network.tables->set_encryptor(encryptor);
-  bft_network.tables->set_history(history);
-  auto backup_consensus =
-    std::make_shared<kv::test::PrimaryStubConsensus>(ConsensusType::BFT);
-  bft_network.tables->set_consensus(backup_consensus);
-
-  GenesisGenerator g(bft_network, gen_tx);
-  g.create_service();
-  user_id = g.add_user({user_caller});
-  CHECK(gen_tx.commit() == kv::CommitResult::SUCCESS);
 }
 
 TEST_CASE("SignedReq to and from json")
